% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdi.R
\name{fdi}
\alias{fdi}
\title{Calculate the Fire Danger Index (FDI)}
\usage{
fdi(temperature, humidity, wind, days_rain, rain)
}
\arguments{
\item{temperature}{Temperature (°C)}

\item{humidity}{Humidity (\%)}

\item{wind}{Wind speed (km/h)}

\item{days_rain}{Days since last rain}

\item{rain}{Amount of last rain (mm)}
}
\value{
A floating point number.
}
\description{
Calculate the Fire Danger Index (FDI)
}
\examples{
fdi(10, 50, 10, 1, 20)
fdi(40, 30, 30, 15, 5)
}
