% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind-factor.R
\name{wind_factor}
\alias{wind_factor}
\title{Calculate wind factor}
\usage{
wind_factor(wind, burn_index)
}
\arguments{
\item{wind}{Wind speed (km/h).}

\item{burn_index}{Burn Index}
}
\description{
Calculate wind factor
}
